pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ray-studio"
version = "0.1.0"
description = "AI-powered marketing asset generator"
authors = [
  { name = "Ray Studio Team", email = "info@raystudio.ai" },
]
dependencies = [
    "click>=8.0",
    "pillow>=10.0",
    "pydantic>=2.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
ray-studio = "ray_studio.cli:cli"
